+++ Horror Game/Library/PackageCache/com.unity.burst/Editor/BurstInspectorGUI.cs
@@ -10,6 +10,12 @@ using System.Text.RegularExpressions;
 using UnityEditor.IMGUI.Controls;
 using UnityEngine;
 
+#if UNITY_6000_2_OR_NEWER
+using TreeView = UnityEditor.IMGUI.Controls.TreeView<int>;
+using TreeViewItem = UnityEditor.IMGUI.Controls.TreeViewItem<int>;
+using TreeViewState = UnityEditor.IMGUI.Controls.TreeViewState<int>;
+#endif
+
 [assembly: InternalsVisibleTo("Unity.Burst.Editor.Tests")]
 [assembly: InternalsVisibleTo("Unity.Burst.Tester.Editor.Tests")]