+++ Horror Game/Library/PackageCache/com.unity.collections/CHANGELOG.md
@@ -1,13 +1,48 @@
 # Changelog
 
+
+## [2.5.7] - 2025-04-11
+
+### Changed
+
+* Updated the `com.unity.entities` dependency to version `1.3.14`
+* Updated the `com.unity.burst` dependency to version `1.8.19`
+* Updated the `com.unity.nuget.mono-cecil` dependency to version `1.11.5`
+* Updated the `com.unity.test-framework dependency` to version `1.4.6`
+* The minimum supported editor version is now 2022.3.20f1
+
+### Fixed
+
+* UnsafeQueue memory leak due to OnDomainUnload callback being discarded by Burst.
+
+
+## [2.5.3] - 2025-02-17
+
+### Fixed
+* xxHash3.Hash128 when length is exactly 16, 128, or 240 bytes.
+
+
+## [2.5.2] - 2025-01-16
+
+### Added
+
+* Added an explicit Mathematics 1.3.2 dependency
+
+### Fixed
+
+* Fixed incorrect behavior of `UnsafeBitArray.ReadOnly.TestNone`
+
+
 ## [2.5.1] - 2024-09-06
 
 ### Changed
+
 * Updated Burst dependency to version 1.8.17
 * Updated Unity Test Framework dependency to version 1.4.5
 * Updated entities packages dependencies
 
 ### Fixed
+
 * Certain cases would cause an ILPostProcessor to fail, blocking compilation, but no more.